# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: TriReservas API
  description: |-
    TriReservas es una aplicación web para la gestión de alojamientos y reservas, con roles diferenciados para 
    usuarios y anfitriones. Desarrollada con Spring Boot, JPA y Angular, ofrece una solución moderna y eficiente para la
    administración de estancias.
    
    Requisitos Funcionales:
    * [Aquí](https://docs.google.com/spreadsheets/d/14pBo766cm17JRFVoYanh_tZuF3VmeBVc5iNh8Avp4dE/edit?usp=sharing) podrá 
    encontrar un apoyo en hojas de cálculo para la identificación de requisitos funcionales y el progreso del proyecto.

    Autores: 
    * Juan Sebastián Londoño Ramírez [github](https://github.com/Juan3bitas)
    * Angie Carolina Grajales Giraldo [github](https://github.com/angiec-grajalesg)
    * Juan Diego García Nieto [github](https://github.com/Juan-Devgo)
  version: 1.1.1
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
...
servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo
paths:
...
  /api/reservations/{lodging_uuid}:
    get:
      tags:
        - Reservation
      summary: Obtener reservas de un alojamiento específico
      description: Obtiene una lista de reservas asociadas a un alojamiento específico.
      operationId: getReservationsByLodging
      parameters:
        - name: lodging_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID del alojamiento para el cual se están obteniendo las reservas.
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
          description: Filtrar reservas por estado (PENDING, CONFIRMED, CANCELLED, COMPLETED)
        - name: checkIn
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filtrar reservas por fecha de inicio (check-in) en formato AAAA-MM-DD
        - name: checkOut
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filtrar reservas por fecha de fin (check-out) en formato AAAA-MM-DD
      responses:
        '200':
          description: Reservas obtenidas satisfactoriamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Reservas obtenidas satisfactoriamente."
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReservationDTO'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "RYA13-1", message: "No autorizado para esta acción. (JWT ausente)"}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "RYA13-2", message: "Prohibido para esta acción (JWT inválido o sin permisos)"}
        '404':
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Alojamiento no encontrado."
                  content:
                    default: {code: "RYA13", message: "No se encontró el alojamiento especificado para la reserva."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-19", message: "Error inesperado al tratar de obtener las reservas."}
      security:
        - bearerAuth: []
    post:
      tags:
        - Reservation
      summary: Crear una nueva reserva
      description: Crea una nueva reserva para un alojamiento específico.
      operationId: createReservation
      parameters:
        - name: lodging_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID del alojamiento para el cual se está creando la reserva.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationDTO'
      responses:
        '201':
          description: Reserva creada satisfactoriamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Reserva creada satisfactoriamente."
                  content:
                    default: null
        '400':
          description: Error en la solicitud
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error en la solicitud."
                  content:
                    default: {code: "RYA03-1", message: "Error en la solicitud al crear una reserva."}
        '404':
          description: Alojamiento no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Alojamiento no encontrado."
                  content:
                    default: {code: "RYA03-2", message: "No se encontró el alojamiento especificado para la reserva."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-20", message: "Error inesperado al tratar de crear una reserva."}
      security:
        - bearerAuth: []
  /api/reservations/{user_uuid}:
    get:
      tags:
        - Reservation
      summary: Obtener reservas de un usuario específico
      description: Obtiene una lista de reservas asociadas a un usuario específico.
      operationId: getReservationsByUserUUID
      parameters:
        - name: user_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID del usuario para el cual se están obteniendo las reservas.
      responses:
        '200':
          description: Reservas obtenidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Reservas obtenidas satisfactoriamente."
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReservationDTO'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "RYA06-1", message: "No autorizado para esta acción. (JWT ausente)"}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "RYA06-2", message: "Prohibido para esta acción (JWT inválido o sin permisos)"}
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Usuario no encontrado."
                  content:
                    default: {code: "RYA06-3", message: "No se encontró el usuario especificado para las reserva."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-21", message: "Error inesperado al tratar de obtener las reservas."}
      security:
        - bearerAuth: []
  /api/reservations/{reservation_uuid}/state:
    patch:
      tags:
        - Reservation
      summary: Actualizar el estado de una reserva existente
      description: Actualiza el estado de una reserva existente identificada por su UUID.
      operationId: cancelReservation
      parameters:
        - name: reservation_uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID de la reserva que se desea actualizar.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reservationState:
                  type: string
                  enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
              required:
                - reservationState
      responses:
        '200':
          description: Estado de reserva actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: false
                  message:
                    type: string
                    default: "Estado de reserva actualizado satisfactoriamente."
                  content:
                    default: null
        '400':
          description: Error en la solicitud
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error en la solicitud."
                  content:
                    default: {code: "RYA04-RYA14-GTR06-1", message: "Error en la solicitud al tratar de actualizar el estado de la reserva."}
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "No autorizado."
                  content:
                    default: {code: "RYA04-RYA14-GTR06-2", message: "No autorizado para esta acción. (JWT ausente)"}
        '403':
          description: Prohibido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Prohibido."
                  content:
                    default: {code: "RYA04-RYA14-GTR06-3", message: "Prohibido para esta acción (JWT inválido o sin permisos)"}
        '404':
          description: Reserva no encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Reserva no encontrada."
                  content:
                    default: {code: "RYA04-RYA14-GTR06-4", message: "No se encontró la reserva especificada para actualizar su estado."}
        default:
          description: Error inesperado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: boolean
                    default: true
                  message:
                    type: string
                    default: "Error inesperado."
                  content:
                    default: {code: "ERRIN-22", message: "Error inesperado al tratar de actualizar el estado de la reserva."}
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    ReservationDTO:
      description: Esquema que representa una reserva para transferencia de datos.
      type: object
      properties:
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        guests:
          type: integer
          format: int8
      required:
        - lodgingUUID
        - checkIn
        - checkOut
        - guests
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT